requirements.txt
flask==2.0.2
kafka-python==2.0.2
